requirements.txt
# Core
streamlit>=1.32.0

# CV parsing
pymupdf>=1.23.0
python-docx>=1.1.0

# NLP / ML
scikit-learn>=1.4.0
numpy>=1.26.0

# Optional: better semantic embeddings (installs separately if GPU is available)
# sentence-transformers>=2.7.0

# HTTP
requests>=2.31.0

# Environment config
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
